+++ src/main/java/com/liga/internship/client/bot/handler/FavoritesHandler.java
@@ -32,9 +32,6 @@ public class FavoritesHandler implements InputCallbackHandler, InputMessageHandl
     private final V1RestService v1RestService;
     private final FavoritesService favoritesService;
     private final ImageCreatorService imageCreatorService;
-    private final FavoritesDataCache favoritesDataCache;
-    private final MainMenuService mainMenuService;
-    private final TranslateToOldSlavService translateToOldSlavService;
     private final TextService textService;
 
     @Override
@@ -89,7 +86,7 @@ public class FavoritesHandler implements InputCallbackHandler, InputMessageHandl
                 status = CAPTION_FAVORITE_FEMALE;
             }
         }
-        String transformedUsername = translateToOldSlavService.translateTextToOldSlav(nextUser.getUsername(),"");
+        String transformedUsername = textService.translateTextIntoSlavOld(nextUser.getUsername());
         return String.format("%s, %s, %s", gender, transformedUsername, status);
     }
