+++ src/main/java/com/liga/internship/client/commons/TextMessage.java
@@ -14,6 +14,7 @@ public class TextMessage {
     public static final String MESSAGE_EMPTY = "Пустота то какая!! Пустота!";
     public static final String MESSAGE_COMEBACK = "С возвращением";
     public static final String MESSAGE_COMEBACK_LATER = "Отличная работа, приходите позже!";
+    public static final String MESSAGE_MANY_WORDS = "Это не твитеръ, лимит 300 букв";
     public static final String CAPTION_LOVE = "Взаимность";
     public static final String CAPTION_FAVORITE = "Любим вами";
     public static final String CAPTION_FAVORITE_FEMALE = "Любима вами";
