+++ src/main/java/com/liga/internship/client/service/TinderService.java
@@ -19,7 +19,7 @@ import java.util.List;
 /**
  * Сервис хендлера голосования предоставляет создание ответов в виде:
  * -текстовое сообщение с главным меню
- * -фото сообщение с главиатурой голосования
+ * -фото сообщение с клавиатурой голосования
  * -изменяемое фотосообщение с клавиатурой голосования
  */
 @Service
